third_party/blink/web_tests/external/wpt/css/selectors/selectors-4/lang-singleton-subtag-matching-expected.txt

This is a testharness.js-based test.
[FAIL] :lang(fr-xenomorph) only matches lang="fr-xenomorph" (not lang="fr-x-xenomorph")
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
